+++ CHANGELOG.md
@@ -27,7 +27,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 
 - Fixed block picking behavior.
 - Fixed fence gate opening in server-side.
-- Fixed quartz blocks behavior.
+- Fixed quartz and purpur blocks behavior.
 - Fixed `BlockFace.UP` axis.
 
 ## 0.4.1 (API 0.7.0) - 2025-5-11